EntityHubRank ist ein Float, und kein Double

+++ enhancement-engines/entities-filter/src/main/java/de/unidue/stanbol/entities/filter/data/EntitiesFilterEngineConfiguration.java
@@ -4,7 +4,7 @@ public class EntitiesFilterEngineConfiguration {
     public static final String MINIMAL_ENTITY_HUB_RANK_PROPERTY = "de.unidue.stanbol.entities.filter.minEntityHubRank";
     public static final String MINIMAL_CONFIDENCE_PROPERTY = "de.unidue.stanbol.entities.filter.minConfidence";
 
-    private double minEntityHubRank;
+    private float minEntityHubRank;
     private double minConfidence;
 
     public double getMinEntityHubRank() {
